fixing some bug

+++ app/src/main/java/com/project/fishbud/ui/main_ui/marketplace/cart/CartAdapter.kt
@@ -49,7 +49,7 @@ class CartAdapter(private var mCallback: CartHarga) : RecyclerView.Adapter<CartA
         fun bind(data: IkanEntity, position: Int) {
             with(binding) {
                 namaIkan.text = data.namaIkan
-                hargaKg.text = formatRupiah(data.harga.toLong(), 1)
+                hargaKg.text = formatRupiah(data.harga, 1)
                 tokoIkan.text = data.tokoIkan
                 Glide.with(itemView.context)
                     .load(data.linkImage)
@@ -59,8 +59,8 @@ class CartAdapter(private var mCallback: CartHarga) : RecyclerView.Adapter<CartA
                     if (countQuantity.text.toString() != "0") {
                         val numb = countQuantity.text.toString().toInt() - 1
                         countQuantity.setText(numb.toString())
-                        hargaIkan.text = formatRupiah((data.harga * numb).toLong(), 0)
-                        val price = (data.harga * numb).toLong()
+                        hargaIkan.text = formatRupiah((data.harga * numb), 0)
+                        val price = (data.harga * numb)
 
                         val tmp = formatRupiah(totalHarga, 0)
                         Log.i("cek_total", "totalHarga paling fix : $tmp")
@@ -74,8 +74,8 @@ class CartAdapter(private var mCallback: CartHarga) : RecyclerView.Adapter<CartA
                     val numb = countQuantity.text.toString().toInt() + 1
                     if ((data.harga * numb) < Int.MAX_VALUE && (data.harga * numb) > 0) {
                         countQuantity.setText(numb.toString())
-                        hargaIkan.text = formatRupiah((data.harga * numb).toLong(), 0)
-                        val price = (data.harga * numb).toLong()
+                        hargaIkan.text = formatRupiah((data.harga * numb), 0)
+                        val price = (data.harga * numb)
                         Log.i("cek_total", "totalHarga: $totalHarga")
                         //interface
                         mCallback.totalHarga(price, position)
